backend/requirements.txt
alembic==1.16.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
filelock==3.18.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
greenlet==3.2.3
hf-xet==1.1.3
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.22.0
onnxslim==0.1.57
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==10.2.0
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
supervision==0.25.1
sympy==1.14.0
tabulate==0.9.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
